project/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
scikit-learn>=1.3.0
xgboost>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
python-dotenv==1.0.0
werkzeug>=3.0.0
imbalanced-learn>=0.12.0
matplotlib>=3.8.0
